ViT/requirements.txt
streamlit
opencv-python-headless
transformers
pillow
ffmpeg